requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
redis>=4.0.2
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
pydantic>=1.8.2
loguru>=0.5.3
tenacity>=8.0.1
python-jose>=3.3.0
passlib>=1.7.4